updated tests for s3 capabilities, added indexes on comment_reference ids and types
